Many many many hours of goblinry later, I've given up on getting the UWP targets working.  At least that means I can free up the 11 gigs (!!!!!!) of useless junk I had to install for universal whatever.
I thought I might try to publish nugets with the 2 small fixes but it seems the way the versioning works, I'd need to publish the full set of stuff.

This repos value as a submodule is dubious at best, but I think that's the way I'll use it for now.